Updated lab07 for testing
Some misspellings were fixed (such as xvbd) and some clarifications were made in Lab 07.  Refresh to see changes.  Biggest change: labeling the filesystem has been removed, but a question was added to Part 3 (2. Use blkid to view information of the partition on xvdb).  The template has been updated to include an answer.